refactor: move audio and query infrastructure into lib
