+++ frontend/src/context/RootProvider.tsx
@@ -2,7 +2,7 @@ import { type ReactNode } from "react";
 import { ThemeProvider } from "./ThemeContext";
 import { AuthProvider } from "@/features/auth";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/shared/query/queryClient";
+import { queryClient } from "@/lib/query/queryClient";
 import { TooltipProvider } from "@/components/ui/tooltip";
 
 export const RootProvider = ({ children }: { children: ReactNode }) => {
