+++ frontend/src/lib/audio/services/analyzer.ts
