+++ frontend/src/lib/audio/services/transcriber.ts
